+++ redis/src/cmd_wrapper.rs
@@ -1,6 +1,7 @@
 use std::ops::{Deref, DerefMut};
 
 use redis::{FromRedisValue, RedisResult, ToRedisArgs};
+use tokio_compat_02::FutureExt;
 
 use crate::ConnectionWrapper;
 
@@ -37,11 +38,11 @@ impl Cmd {
         &self,
         conn: &mut ConnectionWrapper,
     ) -> RedisResult<T> {
-        self.cmd.query_async(DerefMut::deref_mut(conn)).await
+        self.cmd.query_async(DerefMut::deref_mut(conn)).compat().await
     }
     /// See [redis::Cmd::execute_async](https://docs.rs/redis/latest/redis/struct.Cmd.html#method.execute_async)
     pub async fn execute_async(&self, con: &mut ConnectionWrapper) -> RedisResult<()> {
-        self.query_async::<redis::Value>(con).await?;
+        self.query_async::<redis::Value>(con).compat().await?;
         Ok(())
     }
 }