+++ redis/src/lib.rs
@@ -116,6 +116,7 @@ use async_trait::async_trait;
 use redis::{
     aio::Connection as RedisConnection, Client, IntoConnectionInfo, RedisError, RedisResult,
 };
+use tokio_compat_02::FutureExt;
 
 /// A type alias for using `deadpool::Pool` with `redis`
 pub type Pool = deadpool::managed::Pool<ConnectionWrapper, RedisError>;
@@ -176,7 +177,7 @@ impl Manager {
 #[async_trait]
 impl deadpool::managed::Manager<ConnectionWrapper, RedisError> for Manager {
     async fn create(&self) -> Result<ConnectionWrapper, RedisError> {
-        let conn = self.client.get_async_connection().await?;
+        let conn = self.client.get_async_connection().compat().await?;
         Ok(ConnectionWrapper { conn })
     }